Create README.md
EDA_Bank_Case_Study
In this case study ,the company wants to understand the driving factors (or driver variables) behind loan default, i.e. the variables which are strong indicators of default.

Business Understanding : The loan providing companies find it hard to give loans to the people due to their insufficient or non-existent credit history. Because of that, some consumers use it as their advantage by becoming a defaulter. Suppose you work for a consumer finance company which specialises in lending various types of loans to urban customers. You have to use EDA to analyse the patterns present in the data. This will ensure that the applicants capable of repaying the loan are not rejected. When the company receives a loan application, the company has to decide for loan approval based on the applicant’s profile.

Business Objectives : The company wants to understand the driving factors (or driver variables) behind loan default, i.e. the variables which are strong indicators of default. The company can utilise this knowledge for its portfolio and risk assessment.

Download the dataset from the link given as below. https://drive.google.com/open?id=16RQztUqCfJOlbooHqYlJrp6Q7iL65uZB

Sure. Here is some sample code in Python that can be used to apply Association Rule Mining in order to identify relationships between independent continuous variables and dependent continuous variables: